changing how barcode generated

+++ build.sh
@@ -1,25 +1,7 @@
 #!/bin/bash
 
-# make the ISBN barcode
-if [ -f assets/isbn.tex ]; then
-  mv appendix/changelog.tex appendix/changelog.txt
-  if [ ! -f assets/isbn.pdf ]; then
-    cd assets
-    /usr/local/texlive/2021/bin/x86_64-linux/latex isbn.tex
-    if [ $? != 0 ]; then
-      echo "ISBN LaTeX error, exiting"
-      cd ..
-      [ -f appendix/changelog.txt ] && mv appendix/changelog.txt appendix/changelog.tex
-      exit 1
-    fi
-    /usr/local/texlive/2021/bin/x86_64-linux/dvips -E -T 1.85in,1.15in isbn.dvi
-    mv isbn.ps isbn.eps
-    /usr/bin/ps2pdf isbn.eps isbn.pdf
-    rm -f isbn.dvi
-    rm -f isbn.eps
-    cd ../
-  fi
-fi
+# test for canonical
+[ -f canonical.tmp ] && mv appendix/changelog.tex appendix/changelog.txt
 
 # test the LaTeX
 /usr/local/texlive/2021/bin/x86_64-linux/pdflatex SocialistHousing.tex
@@ -34,7 +16,7 @@ fi
 /usr/local/texlive/2021/bin/x86_64-linux/pdflatex SocialistHousing.tex
 
 # make print version
-sed -e s?"digsig\.sty"?"FUBARdigsig.sty"?g < SocialistHousing.tex > SocialistHousing-Print.tex
+sed -e s?"^\\\newcommand{\\\versionforprint}{no}"?"\\\newcommand{\\\versionforprint}{yes}"? < SocialistHousing.tex > SocialistHousing-Print.tex
 
 # three runs
 /usr/local/texlive/2021/bin/x86_64-linux/pdflatex SocialistHousing-Print.tex
